исправляет функцию зума для фото
